Fixed Fortran ASCII to be more symetrical (#443)
I use onefetch a lot, and I work with a lot of FORTRAN code, and the off center rocket just kills me every single time. It was the 3rd result when looking up ASCII rocket, no changes made. This fixes up the rocket and makes it much more pleasing to look at.