Remap White ANSI color to Default
Allows the "white" color to have the greatest chance of being a dark,
visible color for light terminal themes.

A macro is used to remap `White` to `Default` to minimize size of the
diff and prevent contributors from needing to be familiar with the
`Default` color on top of the 8 standard colors.

In other places where `White` was assumed to be the default, `Default`
is now used instead, also.

For #33
Closes #611