Result of `cargo changelog --write`
A sample for non-destructive changelog generation that uses
the git-history as a guide.

It can be filled in manually but will pre-fill itself using
git-conventional commit messages.

It also collects all kinds of interesting information, and
can be customized to not include certain auto-generated sections.

What do you think?

You can play around with it yourself using `cargo install
cargo-smart-release` (a gitoxide project).