Make sure the fastest available zlib configuration is used in `gitoxide`.
To explain: `git-repository` is using the `max-performance-safe` feature
because there are compile issues with `sha1-asm` on windows-msvc which
thus far was elusive to a fix via cargo manifests, which would be pulled
in with `max-performance`.

However, doing so recently also excluded the `zlib-ng-compat` feature
as it has side-effects if `git2` is in the same tree, making
decompressions of objects spuriously fail there during fetches.
Fortunately, in `onefetch` that has never been a problem as it doesn't
fetch (despite the name ;)).

Adding `git-features` as a dependency for the purpose of adding features
makes re-enabling of this performance feature possible. This comes at
the cost of now having to update two crate versions, instead of just
one, to keep them in sync.

Fortunately that's the only issue with this, and it's worth considering
adding another `max-performance-safe-with-zlib-ng` feature to
compensate. The reason it doesn't exist yet is it's lack of scalability:
there are many different zlib configurations, and we don't want to
expose them all in `git-repository`, now or in future. However, common
scenarios should still be facilitated by `git-repository` features, and
maybe this is one of them.