Fix Markdown / Jupyter markup not getting counted (#937)
* Conditionally count lines based on language type

This allows an optional definition of the line type for a language. If
it is defined, it will sum them together. Defaults to using `.code`.

Fixes #933

* Simplify name of LOC counting function

"compile" is a bit of a misnomer, since it's related to any compilation,
but merely counts the number of lines of code for a language.

Co-authored-by: Ossama Hjaji <ossama-hjaji@live.fr>

* Reduce clutter of output `loc` function

Co-authored-by: Ossama Hjaji <ossama-hjaji@live.fr>

* Use templated `loc` function for total LOC

* Remove reference to tokei from languages file

This references internal logic, and is unnecessarily wordy. Instead,
this field is now described in CONTRIBUTING to explain what it is and
what values it can take.

Co-authored-by: Ossama Hjaji <ossama-hjaji@live.fr>

* Add unit test for `get_total_loc`

* Recursively count code stats LOC

* Unset line types for Jupyter Notebooks

All lines are counted from the language's children.

Co-authored-by: Ossama Hjaji <ossama-hjaji@live.fr>

* Remove unnecessary recursion

`.summarise()` accomplishes what recursion was doing.

Co-authored-by: Ossama Hjaji <ossama-hjaji@live.fr>

* Refactor to make templated functions private

This adds public wrapper functions to the templated functions and makes
them private. This makes it so as much code is in actual Rust code, and
leaves the template to only manage the repetitive code.

* Add Jupyter total LOC test

Co-authored-by: Ossama Hjaji <ossama-hjaji@live.fr>