upgrade `git-repository` 0.30 to `gix` 0.36
This upgrade is mainly to benefit from the new names of all `git-*`
crates which now share the `gix-` prefix, with `gix` naturally being
what was formerly `git-repository`.

Now there is no need anymore to rename crates in order to get the `git`
namespace, instead we refer to `gitoxide` as `gix` from the get-go.