Increase performance by decoding the commit only once
Previously, each time we would query commit information, the commit would
lazily be decoded until the point of interest under the hood.

Now we decode everything once, which is faster than what happened before.

Note that diffs are still causing the parents to be decoded even we *could*
pass them in, but it's not worth the complexity for just 100 commits
(the default value for churn).