Avoid exhaustive memory consumption by sending the commit-id instead of its buffer.
The delta-processing happens by referring to a commit, and previously we could
send the whole commit buffer (which is expensive) as the overall amount of buffers
in flight would be bounded.

Now that the bound was removed, it's necessary to limit the costs of the commit,
and we do this by referring to it by id instead.

That way, on the linux kernel, we get these values for memory consumption:

* bounded: 960MB
* unbounded buffer: 2156
* unbounded id: 1033