Assure that `changes()` don't track rewrites which accesses blobs; make filter=blob:none work.
We reproduce the issue by triggering rename detection, which means a change happens on top of a rename.
This tries to access blobs, and that doesn't always work.

It's a good idea to disable rewrite tracking and an oversight in the initial implementation, and one might
say it's tricky given that `gitoxide` automatically does what git does, which may be surprising sometimes
and be a hidden cost even if it doesn't fail.